% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-pheno.R
\name{allcorrect}
\alias{allcorrect}
\title{description Correct phenotypes by the mean of all plants in each growth chamber in each Experiment}
\usage{
allcorrect(dat, pheno, classifier, lineid, op = "trans")
}
\arguments{
\item{dat}{is a dataframe in long format}

\item{pheno}{is a character vector of phenotypes}

\item{classifier}{is a character vector of classifying columns in the dataframe (exp, facility, etc)}

\item{lineid}{the name of the column that contains Accessions (e.g. SALK lines or CS numbers)}

\item{op}{which operation to perform. "trans" means translate by the comparison mean. anything else means scale by comparison mean}
}

