% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-pheno.R
\name{colcorrect}
\alias{colcorrect}
\title{Correct phenotypes by the mean of col plants grown in each growth chamber in each Experiment}
\usage{
colcorrect(dat, pheno, classifier, lineid = "line")
}
\arguments{
\item{dat}{is a dataframe in long format}

\item{pheno}{is a character vector of phenotypes}

\item{classifier}{is a character vector of classifying columns in the dataframe (exp, facility, etc)}

\item{lineid}{the name of the column that contains Accessions (e.g. SALK lines or CS numbers)}
}

